% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_footprint_functions.R
\name{reduce_all_and_remove}
\alias{reduce_all_and_remove}
\title{reduce all output and remove}
\usage{
reduce_all_and_remove(directory)
}
\description{
reduces all output files in directory and
then removes all except the last.  Assumes the output file will be
labelled as "output_<iteration>.RDS" (the default from AnnotFA)
and the reduced files will be
labelled as "less_output_<iteration>.RDS".
}
\details{
@param directory : path to directory.
}
