% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_footprint_functions.R
\name{reduce_all_output_files}
\alias{reduce_all_output_files}
\title{reduce all output files in a directory.}
\usage{
reduce_all_output_files(directory)
}
\arguments{
\item{directory}{path to directory in which files are to be reduced.}
}
\description{
For each AnnotFA output file in the provided directory,
\code{reduce_all_output_files} creates a smaller file with filename
prefixed by "less_" containing the essential parts of the output of
AnnotFA.
}
