% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_footprint_functions.R
\name{reduce_output}
\alias{reduce_output}
\title{reduce output file from AnnotFA}
\usage{
reduce_output(filename)
}
\arguments{
\item{filename}{path to RDS file containing the output from AnnotFA,
which is usually of the form output_<num_iters>.RDS.}
}
\description{
\code{reduce_output} creates a smaller file with filename prefixed by "less_"
containing the essential parts of the output of AnnotFA.
}
