% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_footprint_functions.R
\name{rm_output_files_except_last}
\alias{rm_output_files_except_last}
\title{remove all except last output file.}
\usage{
rm_output_files_except_last(directory)
}
\arguments{
\item{directory}{path to directory.}
}
\description{
remove all AnnotFA output files from a
directory except the last.
}
